+++ cli/tests/test_duplicate_command.rs
@@ -71,8 +71,9 @@ fn test_duplicate() {
     ");
 
     let output = work_dir.run_jj(["undo"]);
-    insta::assert_snapshot!(output, @r"
+    insta::assert_snapshot!(output, @"
     ------- stderr -------
+    Undid operation: 76386e059136 (2001-02-03 08:05:17) duplicate 1 commit(s)
     Restored to operation: 9fdb56d75a27 (2001-02-03 08:05:13) create bookmark c pointing to commit 387b928721d9f2efff819ccce81868f32537d71f
     [EOF]
     ");
@@ -2345,8 +2346,9 @@ fn test_undo_after_duplicate() {
     ");
 
     let output = work_dir.run_jj(["undo"]);
-    insta::assert_snapshot!(output, @r"
+    insta::assert_snapshot!(output, @"
     ------- stderr -------
+    Undid operation: 87bf87a7f06f (2001-02-03 08:05:11) duplicate 1 commit(s)
     Restored to operation: 73a36404358e (2001-02-03 08:05:09) create bookmark a pointing to commit 7d980be7a1d499e4d316ab4c01242885032f7eaf
     [EOF]
     ");