+++ cli/tests/test_git_colocated.rs
@@ -1040,8 +1040,9 @@ fn test_git_colocated_undo_head_move() {
 
     // HEAD should be moved back
     let output = work_dir.run_jj(["undo"]);
-    insta::assert_snapshot!(output, @r"
+    insta::assert_snapshot!(output, @"
     ------- stderr -------
+    Undid operation: 65d1097d37a5 (2001-02-03 08:05:15) new empty commit
     Restored to operation: b528a8c9176f (2001-02-03 08:05:14) new empty commit
     Working copy  (@) now at: vruxwmqv 23e6e06a (empty) (no description set)
     Parent commit (@-)      : qpvuntsm e8849ae1 (empty) (no description set)