+++ CHANGELOG.md
@@ -20,6 +20,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   inside containers or when moved together. Existing workspaces with absolute paths
   will continue to work as before.
 
+* `jj undo` now also outputs what operation was undone, in addition to the
+  operation restored to.
+
 ### Fixed bugs
 
 ## [0.38.0] - 2026-02-04